Styled Components & Dynamic Props
Using styled components we can define conditional styling using dynamic props. Instead of using css classes on our components we can define a prop which then we pass as an argument of an arrow function inside the property we want to dynamically change in the styling section of our styled component thus having 2 different values based on condition.